requirements.txt
streamlit>=1.38.0,<1.41.0
python-dotenv==1.0.1
tqdm==4.66.5

google-generativeai==0.8.2

xmlschema==3.4.2
xmltodict==0.13.0
lxml==5.3.0

pandas==2.2.3
numpy==1.26.4
